requirements.txt
torch
torchvision
timm
numpy
pandas
Pillow
scikit-learn
matplotlib
seaborn
PyYAML
tqdm
tensorboard
